WIP snapshot exploration test
